refactor: share tracked lookup subscriptions
